Finish superficial split of utilities.py
